Refactor screens into separate components and add time-based background
